requirements.txt
absl-py==1.4.0
amqp==5.2.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.3
bcrypt==4.1.3
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cycler==0.11.0
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flatbuffers==23.3.3
flower==2.0.1
fonttools==4.37.4
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
humanize==4.9.0
idna==3.4
imbalanced-learn==0.10.1
inflection==0.5.1
itypes==1.2.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
kombu==5.3.6
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.1
ml-dtypes==0.1.0
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
paramiko==3.4.0
patsy==0.5.3
Pillow==9.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.22.3
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.4
PyYAML==6.0.1
redis==5.0.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
typing_extensions==4.5.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.3
wrapt==1.14.1
